[opb_dram]  initial addition of mapping of opb address space to ddr3 address space
Added more probes to determine the actual opb_addr generated by a "read" command
in opb address space.  turns out its always a multiple of 4.  Turns out when
putting this value directly into the ddr3 address space, i get the counter
back perfectly the way JP writes it in simulink.  this is suspicious to me...

regardless to account for this, I take a step to map opb_addr values
to 32-bit words in the 288 ddr3 address space.  this leaves a word in the ddr3
memory unretrieved (the 9th 32 bit word), but this is just the simple mapping
to start with.  For now, translate is called but the outputs are not used.
I'm just setting things up for some testing.