bus_single_port_bram performance, delay options
BRAM implementation added as mask parameter.
Based on this, BRAM implementation is anticipated and fanout
registers arranged for inputs.
Mask parameters added to be able to specify how fanout registers
are implemented. This is useful where there are constants as input
that may be optimised away by using behavioral HDL.