reorder block mod
fanout_latency and n_bits added to mask parameters.

n_bits allows the use of bus_single_port_rams for data buffers (instead of standard single port rams).
This allows the fanout of address and control signals to be controlled. If n_bits is left at 0 (the default),
the previous behaviour is preserved to preserve backwards compatibility.

fanout_latency controls the amount of latency to be used in the controlling
of fanout into these buffers. The default is 0 to preserver script backwards compatibility.