bus_mux n_inputs mask parameter added
The default behaviour of bus_mux is to divide a single bus input into
portions and use a multiplexer to choose between them. If n_inputs > 1
this changes to dividing multiple buses into portions and multiplexing
between them.

This allows the control of fanout on the sel input where
very large multiplexers are required (or where the sel input to every mux
is not the same).