[translate and mem_opb_attach] adjusts translate code to skip addresses by 8
According to the MIG documentation and the ddr3_controller code,
the MIG is set up with BL8 encoding (Burst Length 8) where word
length is 72, so a burst of 8 is in fact 288*2 bits.  The address
should skip by 8 every time, so that for a given address 2 288 bit
words are retrieved from the ddr3, each on a separate clock cycle.
This has been tested and cuts the errors in half relative
to the mistaken skip by 4 addressing.  Errors now occur on 0.03%
of reads, but the errors can be large as they are either 0 or
have a few bits flipped.