Run update_casper_block on some library blocks
I ran into problems (matlab segfaults to be exact) when running "update
diagram" on a model containing the pfb_fir block.  The problem seemed to be
related to the use of the bram_delay block in the taps of the pfb_fir.  I
noticed that some blocks in casper_library_delays and casper_library_pfbs had
little red "link status" arrows instead of black or gray arrows, so I ran
"update_casper_block(gcb)" on each block that contained a block with a red
arrow.

This also restored the use of the edge_detect block instead of the deprecated
posedge block (which seems to have crept back in as part of commit e00eece!).

With these changes in place, I can successfully run "update diagram" on a model
containing the pfb_fir block.  FWIW, this is with Matlab 2012a and Xilinx 14.6.