Backport delay and pfb libraries to Simulink 7.2
These libraries were backported to Simulink 7.2 by doing a "save as" to the
Simulink 7.2 format from Matlab 2012a.  The resultant libraries then led to
"unrecoverable errors" when being used with Matlab2012a.  The fix for this is
to remove the seemingly unnecessary LinkData sections that appeared as part of
the backport.  This commit still contains the LinkData sections.  They will be
removed in a subsequent commit so that their removal will be documented in git
(i.e. viewable via "git diff").