Remove sync_delays from pfb_fir first_tap and tap
The sync_delays in the PFB "first_tap" and "tap" blocks ended up delaying the
sync out signal by an extra 2^(PFBSize-n_inputs)*TotalTaps samples.  This extra
"multiple-of-PFBSize" delay is not necessary since the subsequent FFT doesn't
care which "window" is preceded by a sync pulse so long as the
sync-to-first-input-sample relationship is maintained.

Deleting these unnecessary sync_delays reduces resource usage and makes testing
the PFB easier since the sync-in-to-sync-out delay is minimized while remaining
valid.