reorder snag fix
Previous reorder init mask additions left the number of inputs mask parameter at 0 and
the init script drawing an empty block for this case. Certain scripts do
not change the number of inputs mask parameter leading to an empty block and
script errors. Default block now with old value of '1' for number of inputs which
should be backwards compatible again with these scripts.