Add support for demuxing of ADC16 outputs
When the ADC chips of the ADC16 board are running in single or dual channel
mode (as opposed to quad channel mode), they output their samples in a way that
needs to be rearranged so that they are presented to the fabric in the usual
CASPER demultiplexed order.  Prior to this commit, the ADC16 yellow block did
not do this and, due to the a/b lane distinction being gone by the time the
samples reach the fabric, the user could not reliably accomplish this task
which requires a/b lane awareness.

Technically, this is really multiplexing the a and b lanes differently for the
different modes, but since the samples appear in a demultiplexed manner in dual
and single channel modes, the terminology used to describe it is "demux
support".