Robustified the effects of setting IDCTRL_100.
Now when this variable is set the tools will instantiate *TWO*
IDELAYCTRLs, one with a hacked 100 MHz reference clock and the
other with normal 200 MHz reference clock. The 100 MHz IDELAYCTRL
is then grouped with the IODELAYs of the QDR's while the other
should by default be used for any other IODELAYs.

This should in theory not break any existing designs if the
IDCTRL_100 environment variable is not set but if it is set the
QDR calibration may succeed at certain fabrics clock rates without
breaking calibration of any other core (for example, the DDR3 core).