Final (hopefully) fix to the conditional IDLYCTRL_100 instantiation.
This conditionally instatiates two IDELAY_CTRL instances if the IDCTRL_100
environmental variable is set, one using 100 MHz and another using 200 MHz,
and only the 200 MHz version if the variable is not set. This differs from
before in the following way: when the variable is set both IDELAY_CTRL's have
their IODELAY_GROUP set to either "IODELAY_100" or "IODELAY_200" respectively
while in the the unset case there are no groups at all.