Remove "||" from mask callback text
This was commented out in the software register mask callback text, but
having the "||" in the text caused problems when backporting to older
Simulink file formats that use '|' as a delimiter without providing any
way (apparently) to escape '|' characters that appear in the callback
text.

This commit just deleted the offending (and commented out!) text.