Make 10 GbE debug counters optional
Commit 3ad784a added debug counters, software registers, and and a new
input port to the tenge_v2 block.  This is nice if one wants to have
these debug counters, but it essentially breaks existing designs that
don't already include this additional port.

This commit makes this extra port and the associated counters and
software registers optional.  To utilize these new features, set the
CASPER_TENGBE_V2_DEBUG environment variable to a non-empty string.