Replace use of "quant" function [reapply 78c67dc]
The "quant" function is part of a toolbox, not the base matlab program,
so it's best to avoid using it.

This was originally from commit 78c67dc, but the `round` aspect was lost
in a merge with ska-sa stuff.  This restores the `round` and it matches
the existing whitespace (assuming tab stop of 8).