Make startsg script more generic and robust
This makes the `startsg` script more generic so it will not need to be
tailored to each local installation.  This essentially moves all of the
local configuration information into a user created `startsg.local`
file that is sourced by `startsg`.  The `startsg` script now includes
some checks and validations of these environment variables.

The `startsg.local` file (in fact any `startsg.*` file) is ignored by
git and does not exist in git; it must be created by the user.

The `startsg.local` file should define the following three environment
variables:

    MATLAB_PATH
    XILINX_PATH
    XILINX_PLATFORM

Here is an example `startsg.local` file:

    export MATLAB_PATH=/opt/MATLAB/R2013a
    export XILINX_PATH=/opt/Xilinx/14.7/ISE_DS
    export XILINX_PLATFORM=lin64