Keep sys_clk at 100 MHz when an ADC clocks ROACH2
When the ROACH2 is being clocked from an ADC, the sys_clk should run at
100 MHz.  The implementation introduced in commit e419ce2 tried to run
sys_clk at the fabric rate, which is likely to be faster than necessary
which makes timing unnecessarily harder to meet.  To further complicate
things, the sys_clk MMCM was configured as if the input clock frequency
were the fabric clock, but that is not the case.

This commit restores the previous sys_clk MMCM configuration (i.e.
calculated for a 100 MHz input clock and a 100 MHz output clock).

The sys_clk MMCM might have a similar problem when clocking the fabric
from aux_clk, but this commit does not address that case at all.