Made changes to start sysgen via vivado or ISE
There are now two startup scripts, startsg_ise
and startsg. These are called by the toolflow
depending which backend is being used.
This is a temporary measure, it's pretty dirty. The
way the start scripts currently exist, startup.m is only
really appropriate for ISE, uin vivado, it reimports
stuff which is already loaded