Add support for multiple WB interfaces.
A single module can now have multiple wishbone interfaces with ports
distinguished by a suffix.

Also changed some of the immutability stuff so that whitespace in names
doesn't mess up ability to recognise two ports/signals as the same.
Changed immutability so that attributes equal to None can be
overwritten. This allows 'signal' to be added as a constructor argument
to ports, allowing a signal to be specified positionally, and saving
irritating signal=... arguments for simple port cases.