Add wide_bram_vacc block
The `wide_bram_vacc` block is a vector accumulator based on DSP48 slices and
BRAMs, specifically designed for wide accumulators (48 < output width <= 96).

Like a `bram_vacc`, it outputs its previous accumulated vector after a
`new_acc` pulse is received, but the `wide_bram_vacc` block has one extra cycle
of latency from the `new_acc` pulse to the rising edge of the `valid` output.

The "Data IN" (din) value is used if "Data OK" (dok) is 1, otherwise a 0 value
is accumulated instead.  This can be used, for example, to implement blanking
without requiring external multiplexers.

The block was surgically added to casper_library_accumulators.mdl so this
change should be somewhat mergeable (or at least easy to do surgery with).

A test model (wide_bram_vacc_test.mdl) and a script to run it
(run_wide_bram_vacc_test.m) are also being added.  To run the test model, open
it and then run the run_wide_bram_vacc_test script.  If things work, it will
print "OK"; otherwise it will show an error message.

This replicates commit c72a761, but with reg_retiming turned on for all delay
blocks.