Restore xeng outports to pre-056200 ordering
The xeng_init script uses "reuse_block" to setup or create the outports.  When
the library block had outports (before commit 056200 setup the xeng block to be
empty in the library), reuse_block would reuse the outports and their existing
numbering.  Commit 056200 made the library block empty so then the resue_block
calls created the outports.  Because the reuse_block calls do not assign
outport numbers, the outports get numbered based on their creation order.  As
it turns out, the ordering of the outport reuse_block calls was not the same as
the ordering of the outport numbering when the library block had outports, so
newly created xeng blocks got a different outport ordering than xeng blocks
created from mlib_devel versions prior to commit 056200.

This commit restores the xeng outport ordering to the way it was before commit
056200.

NB: Any models that created an xeng block using mlib_devel version from 056200
to this commit's parent will experience an outport reordering if
"update_casper_block" is called on the existing block.