Make save_state treat library blocks differently
save_state used to save the blocks state (i.e. hash of parameters) even
for library blocks, but this was problematic for library blocks.
Becasue munge_block does not disable the LinkStatus of linked blocks
within a library, the linked library block's implementation would revert
back to the ancestor blocks default implementation and the mask
initialization script (if it utilized same_state/save_state) would not
redraw the implementation to match the mask dialog parameters because
the state saved by save_state matched the dialog parameters!

This commit fixes save_state so that it will set the state field to be
empty when passed a block that resides in a library.  This ensures that
the block's init script will run all the way through the first time it
is placed in a model.