Fix linked sub-block problem in window_delay block
The `window_delay` block in the `casper_library_delays` library contains
two linked `edge detect` blocks, one rising edge sensitive and one
falling edge sensitive.  It had a strange problem where the falling edge
sensitive block was behaving like a rising edge sensitive block.  This
prevented the `window_delay` block from ever outputting a pulse!

In commit e9e3c2c, these blocks had been updated via
`update_casper_block` which ends up calling the blocks' init scripts.
The init scripts did their thing which involves calling `munge_block`
near the beginning and `save_state` at the end.  `munge_block` normally
disables the block's link, but not if the block is in a library, so the
`munge_block` call did nothing.  Then `save_state` saved the state of
the block based on the mask dialog parameters.  This (I think) was the
cause of the problem because the "negedge" `edge detect` block ended up
with its saved state matching the mask dialog, but the link was still
active (because `munge_block` did not disable it) so the modified
contents were ignored/replaced with the contents from the ancestor block
which is configured by default for rising edge.  This default
implementation was never overwritten by the mask init script because
`same_state` always returned true because the saved state matched the
dialog parameters (but not the implementation).

To fix this, `save_state` has been modified to set the `struct.state`
field to `[]` if the block resides in a library.  This means that the
block's init function will always run through to completion (i.e. never
return early because same_state returns true) until it is placed in a
model and `save_state` actually saves the block's state.

It would probably be a good idea to remove any saved state from all
library blocks, but that's a bigger change for another day...