Allow using a symbolic link to startsg
Users can now create a `startsg` symlink in a directory that is on their
`PATH` and point it to the `startsg` script in their `mlib_devel`
directory.  This makes it easier to run `startsg` from within a project
directory instead of the `mlib_devel` directory.

The `startsg` script looks for `startsg.local` in the same directory
that the script lives in.  In the past, this was done by looking in the
directory part of `$0`.  That works fine if `startsg` is invoked from an
`mlib_devel` directory, but not if `startsg` is a symlink in a
non-`mlib_devel` directory pointing to the "real" `startsg` script that
is in an `mlib_devel` directory.

The new `startsg` script uses the `readlink` command to determine the
"real" `startsg` script and from there it determines `mlib_devel`
correctly even if the user executed `startsg` via a symlink.  The
`readlink` command is part of the `coreutils` package on Debian/Ubuntu
and RedHat/CentOS/SL so I think it is safe to assume that is will exist
on the user's system.