Remove erroneous `which` call from gen_xps_mod_ucf
The `which` function in matlab looks up the location of functions,
variables, models, java classes etc.  In `gen_xps_mod_ucf`, it was being
used in an apprarent attempt to lookup UCF files in a given directory,
but that is not an appropriate use of `which` and it is not necessary,
so the call to `which` was simply removed.