Remove unused/unneeded decode from sys_block.v
The sys_block pcore provides the sys_clk_counter value that software can
sample at word offset 4 to estimate the clock frequency.  The value at
word offset 4 is a registered version of the counter.  The sys_block
also provided access to an UNregistered version of the same counter.
This unregistered version was not exposed in core_info.tab yet it
creates a source of potential timing problems so access to it is being
removed here in this commit.