Change QDR latencies -- calibrates at 312MHz
There was something fishy about the QDR controller.
It wouldn't calibrate over ~250MHz. On looking at the
errors (values read vs. values written) produced by
the calibration script, it didn't seem like bit errors.
There were off-by-ones in addresses of values read vs written
and patterns of addresses that were being read as 0.

This commit changes the latency by half a cycle, and has
been tested to calibrate OK at 312.5 MHz (2500MHz on adc5g).
It also seems to work at 200 MHz. However,

No stress testing of the QDR has been performed to see how
good this calibration is, nor has it been tested on multiple boards
or QDR chips yet.

Since the IODELAY range is usually than an fpga clock cycle,
really what we need is the ability to dynamically insert half clock
cycle delays into the data path, to cope with a wide range of clock frequencies.
(The ROACH1 interface could do this)
With this in mind, this commit doesn't really fix anything, it just
brings the zero-delay point of the IODELAY block to a point that
is reasonable for the QDR traces at 312.5MHz