Added a statement to turn on parallel synthesis in system.tcl
This causes the XST step in the XPS build process to synthesize cores
in up to four different processes. For large CASPER designs with a lot
of yellow blocks this can speed up the compile process quite a bit.

NOTE: this is known to work in 14.7 but may not work in earlier versions.
In any event it should not stop the build process if it does not work.