Enabled multithreading in map and par.
For map this means it will run in up to 2 process and for par it will
run in up to 4. This has been shown to significantly speed up compiles
for very large designs.

NOTE: known to work in 14.7 but may not work in earlier versions.
Hopefully if the option is not supported map/par will simply ignore the
option but I do not know this for sure.