Add support for yellow block children
The YellowBlock class now has the method gen_children.
This method returns [] unless overridden. If overridden,
it should return a list of new YellowBlock instances which the
parent wants to spawn. The toolflow now calls this method
recursively over yellow block instances until noone wants
to procreate ay further.

This commit also changes the instance counter so that the first
instance gets an id of 0 (used to be 1). A new attribute
'i_am_the_first' is added, as a shortcut to inst_id==0