Support est_brd_clk() command
This commit changes the sys_block hdl to include a clock counter register.
It adds the address of this register to snap's base core_info.tab
with the name sys_clkcounter, so that corrs est_brd_clk() method
will work as it does with ROACH.

The addition of the clock counter register to the sys_block had to
be propagated up to snaps top.v, which now has a declaration of
user_clk in the base top verilog file.
This has the knock on effect that yellow blocks must be prevented
from trying to add their own user_clk signals to top.v

Hence this commit appears more complicated than one might expect.
The xsg block is changed so id doesn't add user_clk to top, but just
assigns user_clk{90,180,270} with the appropriate source.
Other yellow blocks were altered so that user_clk ports are added to
instances with the parent_sig=False argument, so they don't get
automatically re-declared.