Allow software override of QDR simulink write i/f
QDR calibration requires that the qdr software interface
be able to write to the QDR, then readback values to check
for glitches.
By default the calibration routine (in this repo) writes to the
top half of the QDR for these tests. If the simulink interface
is not using the top half of the QDR this is fine, but if it is,
writes from the simulink interface will clobber the software writes
and cause calibration to fail.
This commit adds the ability for the qdr software controller
to disable the simuilink write interface. This allows calibration
to succeed even if the simulink interface is using the whole QDR
and has no simulink-side logic for disabling writes.