Fix location of casper_create_ppr output directory
The output location of casper_create_ppr.tcl was in the
`XPS_ROACH2_base` directory itself.  Unfortunately, this directory gets
removed and recreated by the "Copy Base Package" step of the casper_xps
toolflow.  This changes the script so that the created planahead project
will be in a directory that is a sibling to the XPS_ROACH2_base.

Here is a partial directory tree before this fix:

    +- project_dir/
       +- my_model.mdl
       +- my_model/
          + XPS_ROACH2_base/
            +- planahead/
               +- my_model.ppr

Here is a partial directory tree after this fix:

    +- project_dir/
       +- my_model.mdl
       +- my_model/
          + XPS_ROACH2_base/
          +- planahead/
             +- my_model.ppr