Terminated outputs and break p_to_s links
The parallel to serial block no longer exists in the main casper
repo. This commit removes the p_to_s masks in the blocks
being used in the DAC blocks. This means that they can't be updated
to different settings, but will update and compile fine in the configuration
required by the DAC.

This commit also ran xlAddTerms inside the yellow blocks to terminate
gateways and prevent simulink warnings.

NOTE: the current casper-astro mlib_devel doesn't have the pcores for the
dac_mkid_4x -- so this block won't work anyway until these are added.