Update MKID ADC MMCM and clock phase
This is an attempt to improve the MMCM clock parameters for clocking the ADC at >480 MHz.

I also set the IDDR clock to clk90 instead of the default clk. From the timing diagrams in the datasheet, the data transitions on the same edge as the DRDY signal. DRDY is the input to the MMCM and clk is in phase with DRDY. We actualy want to sample the data 90 degrees out of phase with DRDY, so I think the IDDR clock should be clk90.
I ran into some glitches using the default clk choice, and when I switched to clk90, they went away, hence this commit.