Print name of BOF file that is created
The name is not known a priori since it contains a build timestamp.
Outputting it upon completion lets the user know the file name so they
don't need to go looking for it.

Some mysteries make life more interesting, but the name of the BOF file
you just built is not one of those mysteries.