Upate extract_entity to support Vivado sysgen
When black boxing a model using System Generator for Vivado, the
compilation target should be set to "Synthesis Checkpoint" rather than
"NGC Netlist" as in previous versions.  This generates HDL and
synthesizes it into a netlist that is saved as a "design checkpoint
file".  The `extract_entity.m` script required two changes to work with
this new version of System Generator.

1. The generated HDL ends up in a slightly different directory
   structure from earlier versions so `extract_entity` now looks in this
   new location if the HDL cannot be found in the other locations that
   it already searches in.

2. Ports with more than one bit are now defined as being `(8-1 downto 0)`
   rather than simply `(7 downto 0)`.  This seems to confuse System
   Generator's "Black Box Wizard" so `extract_entity` does the math so
   it can output the simpler form to make life easier later on.