Add `DCP` block to xps_library and jasper_library
The xps_library now has a `DCP` yellow block that can be used like the
old `pcore` yellow block to include netlists created by "black boxing"
another System Generator model.  The `DCP` block has a single parameter,
`dcp_file`, that currently must be an absolute path to the design
checkpoint (`*.dcp`) file created as the "Synthesis Checkpoint"
compilation target.

One gotcha with current Vivado versions is that the generated netlist
does not have a `ce` port corresponding to the `clk` port.  Without the
matching `ce` port, the Black Box Wizard will not recognize the `clk`
port as a clock port and will assume the entire netlist is
combinational.  To avoid this problem, add a Gateway In block named `ce`
to the model being black boxed and set its type to `UFix_1_0`.