Make startsg and startup.m more versatile
Users can now make a symlink to `startsg`.  `startsg` will look for
local configuration definitions in the current directory in a file named
`startsg.local`.  If it is a symlink, the `startsg` part will be
replaced with the symlink name.  For example, if symlink `mystartsg`
points to `startsg`, then local definitions will be taken from the file
named `mystartsg.local`.  The local definitions file is first searched
for in the current directory, then in the directory where the "real"
`startsg` script lives.  If not found, the deprecated files
`vivado_config.local` and then `vivado.sh` are searched for inthe same
directory where `startsg` lives.

Additionally, startsg now saves the current working directory in an
environment variable named CASPER_STARTUP_DIR.  startup.m looks for this
variable, and if found will cd there.  This means that users can launch
the tools using startsg without having to change back from the
mlib_devel directory to their project directory once Matlab is started
up.

These changes are especially convenient when `mlib_devel` is a git
submodule of an application's git repository.