Make calibration check disable fabric interface
A bug existed whereby the fabric interface was disabled when
searching for a calibration, but re-enabled on applying the
delays prior to checking the calibration.

This commit makes the calibration check explicitly disable
the fabric interface before doing the check.
It's up to the user to re-enable the interface after calling
this method. (The qdr_cal method already does this)