Add support for timer/counter
This sets up the timer/counter as a 64 bit counter that counts CPU clock
cycles (10 ns period).  This provides a way to measure the passage of
time.  Two functions exist to get at the time.  The first is
`read_time()` which returns the 64 bit timestamp from the counter.  The
second is `tick_time()` which returns the middle 32 bits of the 64 bit
timestamp (less resolution and less range than the 64 bit version, but
half the size).