Change errant printf() to xil_printf()
Using printf() results in way too much bloat!  There was one little
printf call that pulled in about 64 KB of code!!!  This was used because
xil_printf does not support %f.  The workaround is to multiply the
floating point temperature by 10 then cast to int.  This results in a
slight rounding error that biases the average down by 0.05 degrees C,
but this is completely acceptable.  Then, print int/10 for the
integer degrees and int%10 for the one decimal place precision.