Cleanup and packet RX improvements
casper_eth.c is greatly cleaned up and the RX code has been improved to
be more alignment tolerant of the pbuf payloads.  The byte swapping of
RX data could be done in the RX buffer if only the RX buffer were not
read-only.  I'm note sure why it has that restriction.  I suspect it is
just an optimization of sorts or just an oversight in the gateware that
could be removed without significant impact to the gateware performance.

Until then, the byte swapping happens into a statically allocated array
of 380 uint32_t elements.  This further restricts the HEAP_SIZE so it
has been reduced accordingly in Makefile.