Enable AUTOIP to use 169.254.x.y IP if DHCP fails
If DHCP fails, we would still like to determine some IP address to use.
Using AUTOIP as a fallback provides this functionality.  If DHCP fails,
the board should self-configure an IP address in the 169.254.0.0/16
subnet.  The default fail-over is after 9 DHCP DISCOVER attempts.  I'm
not sure how long that will take, but it is desirable to keep trying
DHCP for more than 30 seconds since switches with Spanning Tree Protocol
(STP) enabled might prevent access to the network for 30 seconds after
the link comes up.