Add files related to Compact Sorted List
See the comments in csl.c for info about the Compact Sorted List (CSL)
data structure.  The core_info.tab data will be stored in the microblaze
memory as a CSL.

The files `csl.c` and `csl.h` provide the `find_payload()` function
which is used to find CSL payload data corresponding to a given key.

The `test` subdirectory contains `csl_test.c` and `Makefile` for a unit
test program, `csl_test`, that tests the `find_payload()` function.

The `util` subdirectory contains a Ruby script, `cit2bin.rb` that can be
used to convert a CASPER core_info.tab file into a CSL.  To use it:

    cit2bin.rb core_info.tab > core_info.bin