Add -O2 optimization flag to CFLAGS
Compiling with optimization level `-O2` reduces the code size
significantly, which allows the heap size to be increased significantly.

With optimizations enabled, we need to be more diligent about declaring
things `volatile`.  The two things that needed to be modified to get the
currently implemented network related functionality working were the
macros for the TX and RX buffer size pointers.  As more functionality is
added and more testing is done, more such modifications may be
necessary.