Always link core_info or dummy_info
If core_info.tab or core_info.bin exists, it will be used to create
core_info.o, which will be linked to the main executable,
executable.elf.  If neither of those files exist, dummy_info.bin will be
used to create dummy_info.o which will be linked with executable.elf.

core_info.o and dummy_info.o will define the symbol:

    _binary_core_info_bin_start

That symbol will refer to the first character of the core_info (or dummy
core_info)  CSL.  Taking the address of that symbol will return a
pointer to the core_info CSL.

This means that the source code need not be compiled differently
depending on whether a real core_info CSL is linked in.  It also makes
updating the core_info CSL a purely link-time operation.