Change the linking of core_info/dummy_info data
The core_info (or dummy_info) data is now linked in its own `.core_info`
section in `executable.elf`.  This section occupies the top 4KB of
memory.  The first two bytes of this section contain the length of the
core_info data that immediately follow.  The length field is stored in
big endian format.  The symbols `_core_info_size_be` and `_core_info`
are defined by the linker script so code can refer to these data
symbolically.  The linker script uses those symbol names even when the
core_info data really comes from dummy_info.bin.

To access these symbols in C source code, use code like this:

    extern const uint16_t _core_info_size_be;
    extern const uint8_t _core_info;

    uint16_t core_info_size = ntohs(_core_info_size_be);
    uint8_t * core_info_ptr = &_core_info;

The placement at a fixed address is not for the benefit of C programs,
but rather to allow the core_info data to be updated post-linking by
using Xilinx's `updatemem` command.