Give static line_buf file scope, add lidx field
The `line_buf` buffer that was static to the
`casper_tapcp_read_listdev_ascii()` function now has file scope.  This
is in preparation of other read functions with line based ASCII mode
outputs.

The `tapcp_state` structure now have a `lidx` field that is used to keep
track of the next `line_buf` byte to output.  This avoids the need to
use the inapprpriately named `nleft` field for this purpose.