Reduce main loop overhead
Do not call `usleep()` in the main loop as it delays our response time
to incoming packets.

Reduce the verbosity of the periodic heartbeat message since that can
also delay our response time.  This means we longer test for and display
the fact that we are on a little endian system.