Change where platform yellow block is instantiated
It was added to the yellow block list within toolflow.py.
Now it is a child block of xsg.py. This seems a little neater,
since the instantiation of this block can now be traced from the
jasper.per file. Insertion of a special yellow block by the toolflow
which doesn't appear in jasper.per is a bit of unnecessary black magic.