Add typecodes to yellow blocks
These are specified in yellow_blocks/yellow_block_typecodes.py

The toolflow now (in addition to core_info.tab) generates a
core_info.jam.tab file, which incorporates a typecode for each
different type of yellow block.
This is designed to allow a microblaze to figure out what ethernet
interfaces there are in a design so that it can start talking.