Update MB block diagram with 115200 baudrate UART
Add associated elf file for using the Microblaze core
in the new block diagram.

Note:
executable.elf is an executable with a baked-in core_info register map.
executable.mem is the memory contents associated with the elf file, with
the core_info register map stripped out. To use this file in an FPGA,
you need to first append an appropriate .mem file for your core_info.tab,
and then inject to a bitstream using Xilinx's updatemem tool.