Allow tcl commands to be run at different stages
This commit changes the tcl_cmd attribute of the
backend from a list of commands to a dictionary of lists.
The dictionary has entries 'init', 'pre_synth',
'post_synth', 'pre_impl', 'post_impl', 'pre_bitgen',
'post_bitgen'. Commands can be appended to each list,
and will ultimately be executed in the above order.

For compatibility with this change, the yellow_block
method "gen_tcl_cmds()" now returns a dictionary
of lists, matching the above structure.

This change was primarily implemented so that the microblaze
yellow_block could trigger injection of microblaze memory contents
via the updatemem command, which must be run after bitgen.
However, it should enable various bits of code in the Backend class
to be moved into yellow blocks. For example, code in the backend
to generate PROM files (which is hardware specific, owing to
bus-widths / memory types / etc.) can now be moved into the
platform-specific yellow blocks (snap.py, skarab.py, etc)