Always instantiate cpu interface
The original code left out the CPU interface if
the user had selected DISABLE_CPU_TX *or* DISABLE_CPU_RX.
If this happens the build will fail, as the OPB infrastructure
ends up connected to a bunch of undriven input signals.

Further, the "standard" taken from the 10GbE core is that the
CPU interface, which allows setting of the ARP cache, local
MAC/IP/etc is always instantiated. Whether or not the CPU
is allowed to send and receive packets over the Ethernet interfaces
is a different kettle of fish.

This commit always instantiates the CPU interface, regardless
of user settings.