Remove peripherals arg from YellowBlock.gen_constraints
This argument was added to the method to allow blocks
in the skarab library to figure out the name of a user's
40GbE yellow block. This name was used to format async
clock constraint commands.

The blocks requiring these constraints now get the clock
using wildcard matching, which means they don't need to know
the name of the 40GbE block.

The API to YellowBlock has thus been reverted to it's previous
state -- gen_constraints requires no arguments.