move address space of adc5g controller
Move the ADC5g controller address space from
0x20000 - 0x2ffff to 0x10000 - 0x1ffff
so that it can operate in harmony with other ADCs, in the
slightly unusual case where a user wants two different
ZDOK ADCs on a single ROACH.
This fix is a bit lame, because most/all of the ADCs use the
same 0x20000 range for their controllers, so this commit
only attempts to solve the case where one adc is an ADC5g.
For example, an iADC + ADC16 still won't work.

The toolflow doesn't dynamically generate addresses on the fly,
so a proper fix would involve trawling the repo to find unused
address ranges which could be assigned to different kinds of ADC.