Allow demux-2 inputs
The fft-wideband-real block is based on a complex, biplex
fft core, which requires inputs to be multiples of 4 parallel real streams
to FFT.
The block used to check that this requirement was met by demanding that
the demux-factor -- "simultaneous samples per FPGA clock -- was a multiple
of 4. In reality, if multiple independent ADC streams are being transformed,
the requirement need only be N_ADC_STREAMS * DEMUX_FACTOR is a multiple of 4.
This commit implements this more relaxed requirement