Make TFTP timeout 10s
This is probably too long, but it takes several 100ms for a
flash page erase, so the client usually runs with a relatively
long timeout. If the server-side timeout is too short then
if a packet is lost by the time the client resends the server
will have given up on the connection.