Ignore duplicate write packets
the LWIP implementation of TFTP didn't
have any logic to ignore duplicate write packets
(i.e., a write packet resent by a client because
it failed to receive an ack on a previous packet).
This meant that duplicate packets were written twice,
and the write address was erroneously incremented.

This commit adds logic to track the expected TFTP
blknum (i.e., packet number) and re-ack, but not re-write
any duplicated packets.